sane type

+++ src/hashtable.h
@@ -9,7 +9,7 @@ typedef int (*hashtable_cmpfunc)(void *a, void *b, void *userdata);
 
 // should be considered an implementation detail
 struct HashTableItem {
-	unsigned long keyhash;    // longs might be useful for a huuuuuge hashtable
+	unsigned int keyhash;
 	void *key;
 	void *value;
 	struct HashTableItem *next;
@@ -27,14 +27,14 @@ struct HashTable {
 struct HashTable *hashtable_new(hashtable_cmpfunc keycmp);
 
 // return STATUS_OK on success, and on failure STATUS_NOMEM or an error code from cmpfunc
-int hashtable_set(struct HashTable *ht, void *key, unsigned long keyhash, void *value, void *userdata);
+int hashtable_set(struct HashTable *ht, void *key, unsigned int keyhash, void *value, void *userdata);
 
 // return 1 if key was found, 0 if not (*res is not set) or an error code from cmpfunc
-int hashtable_get(struct HashTable *ht, void *key, unsigned long keyhash, void **res, void *userdata);
+int hashtable_get(struct HashTable *ht, void *key, unsigned int keyhash, void **res, void *userdata);
 
 // delete an item from the hashtable and set it to *res, or ignore it if res is NULL
 // returns 1 if the key was found, 0 if not and an error code from cmpfunc on failure
-int hashtable_pop(struct HashTable *ht, void *key, unsigned long keyhash, void **res, void *userdata);
+int hashtable_pop(struct HashTable *ht, void *key, unsigned int keyhash, void **res, void *userdata);
 
 // everything except key and value should be considered implementation details
 struct HashTableIterator {
@@ -69,9 +69,9 @@ void hashtable_fclear(struct HashTable *ht, void (*f)(void*, void*, void*), void
 // deallocate a hashtable, must be cleared first
 void hashtable_free(struct HashTable *ht);
 
-//int hashtable_contains(struct HashTable *ht, void *key, unsigned long keyhash, void *userdata);
+//int hashtable_contains(struct HashTable *ht, void *key, unsigned int keyhash, void *userdata);
 //int hashtable_equals(struct HashTable *ht1, struct HashTable *ht2, hashtable_cmpfunc valuecmp, void *userdata);
 
-//unsigned long hashtable_stringhash(unicode_t *string, size_t stringlen);
+//unsigned int hashtable_stringhash(unicode_t *string, size_t stringlen);
 
 #endif   // MAPPING_H
